+++ src/ClientInterface.php
@@ -8,6 +8,7 @@ use Aceproxies\ResellerApi\Endpoint\BalanceInterface;
 use Aceproxies\ResellerApi\Endpoint\HealthInterface;
 use Aceproxies\ResellerApi\Endpoint\OrdersInterface;
 use Aceproxies\ResellerApi\Endpoint\ProductsInterface;
+use Aceproxies\ResellerApi\Endpoint\ServicesInterface;
 
 interface ClientInterface
 {
@@ -18,6 +19,8 @@ interface ClientInterface
     public function orders(): OrdersInterface;
 
     public function products(): ProductsInterface;
+
+    public function services(): ServicesInterface;
 
     public function getApiVersion(): string;
 }