+++ src/Endpoint/ServicesInterface.php
@@ -1,0 +1,22 @@
+<?php
+
+declare(strict_types=1);
+
+namespace Aceproxies\ResellerApi\Endpoint;
+
+use Aceproxies\ResellerApi\Exception\ApiException;
+use Aceproxies\ResellerApi\Exception\InvalidResponseException;
+use Aceproxies\ResellerApi\Exception\TransportException;
+use Aceproxies\ResellerApi\Response\ServiceListResponse;
+use InvalidArgumentException;
+
+interface ServicesInterface
+{
+    /**
+     * @throws ApiException
+     * @throws InvalidResponseException
+     * @throws TransportException
+     * @throws InvalidArgumentException
+     */
+    public function list(?int $page = null, ?int $limit = null): ServiceListResponse;
+}